third_party/blink/web_tests/flag-specific/highdpi/http/tests/intersection-observer/cross-origin-iframe-with-fixed-position-element-expected.txt

This is a testharness.js-based test.
FAIL IntersectionObserver target in cross-origin iframe with fixed position reports viewport intersectino. assert_equals: Fixed position div intersects viewport.left expected 25 but got 24.65625
Harness: the test ran to completion.
